+++ datafusion/core/src/datasource/empty.rs
@@ -28,7 +28,7 @@ use datafusion_common::project_schema;
 use crate::datasource::{TableProvider, TableType};
 use crate::error::Result;
 use crate::logical_expr::Expr;
-use datafusion_datasource::empty::EmptyExec;
+use datafusion_physical_plan::empty::EmptyExec;
 use datafusion_physical_plan::ExecutionPlan;
 
 /// An empty plan that is useful for testing and generating plans
