+++ datafusion/core/tests/custom_sources_cases/mod.rs
@@ -39,9 +39,9 @@ use datafusion_catalog::Session;
 use datafusion_common::cast::as_primitive_array;
 use datafusion_common::project_schema;
 use datafusion_common::stats::Precision;
-use datafusion_datasource::placeholder_row::PlaceholderRowExec;
 use datafusion_physical_expr::EquivalenceProperties;
 use datafusion_physical_plan::execution_plan::{Boundedness, EmissionType};
+use datafusion_physical_plan::placeholder_row::PlaceholderRowExec;
 use datafusion_physical_plan::PlanProperties;
 
 use async_trait::async_trait;
