+++ datafusion/physical-plan/src/work_table.rs
@@ -20,15 +20,13 @@
 use std::any::Any;
 use std::sync::{Arc, Mutex};
 
+use crate::execution_plan::{Boundedness, EmissionType};
 use crate::memory::MemoryStream;
-use datafusion_physical_plan::execution_plan::{Boundedness, EmissionType};
-use datafusion_physical_plan::{
+use crate::{
     metrics::{ExecutionPlanMetricsSet, MetricsSet},
     SendableRecordBatchStream, Statistics,
 };
-use datafusion_physical_plan::{
-    DisplayAs, DisplayFormatType, ExecutionPlan, PlanProperties,
-};
+use crate::{DisplayAs, DisplayFormatType, ExecutionPlan, PlanProperties};
 
 use arrow::datatypes::SchemaRef;
 use arrow::record_batch::RecordBatch;
